requirements.txt
bottle==0.12.20
Pykka==1.2.1
python-mpd2==3.1.0
